refactor: Resolve TODOs, organize utilities
